+++ weldx/tests/asdf_tests/test_asdf_util.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass
 import numpy as np
 import pytest
 
-from weldx import WeldxFile
+from weldx.asdf.file import WeldxFile
 from weldx.asdf.util import (
     dataclass_serialization_class,
     get_highest_tag_version,